requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
google-generativeai>=0.3.0
python-multipart>=0.0.6

# PPT 生成依赖
python-pptx>=0.6.21

# ppt-master 工具链依赖
PyMuPDF>=1.23.0
Pillow>=9.0.0
numpy>=1.20.0
svglib>=1.5.0
reportlab>=4.0.0

# 多模态支持（可选，需要 Python 3.11/3.12）
# 如果安装失败，可以注释掉以下两行，仅影响图片/视频功能
# pip install rapidocr-onnxruntime faster-whisper
rapidocr-onnxruntime>=1.3.0
faster-whisper>=0.10.0
